Make loiName nullable, move LoiNameDialog to its own file.
