+++ ground/src/main/res/navigation/nav_graph.xml
@@ -89,7 +89,7 @@
       <argument
       android:name="locationOfInterestName"
       type="string"
-      app:nullable="false" />
+      app:nullable="true" />
     <argument
       android:name="jobId"
       type="string" />
@@ -113,7 +113,7 @@
       <argument
         android:name="locationOfInterestName"
         type="string"
-        app:nullable="false" />
+        app:nullable="true" />
       <argument
         android:name="jobId"
         type="string" />
